Add verbose version info output
