+++ Makefile
@@ -1,6 +1,6 @@
 ################################################################################
 
-# This Makefile generated by GoMakeGen 2.1.0 using next command:
+# This Makefile generated by GoMakeGen 2.2.0 using next command:
 # gomakegen --mod .
 #
 # More info: https://kaos.sh/gomakegen
@@ -94,6 +94,6 @@ help: ## Show this info
 		| sed 's/ifdef //' \
 		| awk 'BEGIN {FS = " .*?## "}; {printf "  \033[32m%-14s\033[0m %s\n", $$1, $$2}'
 	@echo -e ''
-	@echo -e '\033[90mGenerated by GoMakeGen 2.1.0\033[0m\n'
+	@echo -e '\033[90mGenerated by GoMakeGen 2.2.0\033[0m\n'
 
 ################################################################################
