+++ README.md
@@ -18,16 +18,10 @@ Yo is a command-line YAML processor.
 
 #### From source
 
-To build the Yo from scratch, make sure you have a working Go 1.17+ workspace (_[instructions](https://golang.org/doc/install)_), then:
+To build the Yo from scratch, make sure you have a working Go 1.18+ workspace (_[instructions](https://golang.org/doc/install)_), then:
 
 ```
-go get github.com/essentialkaos/yo
-```
-
-If you want update Yo to latest stable release, do:
-
-```
-go get -u github.com/essentialkaos/yo
+go install github.com/essentialkaos/yo@latest
 ```
 
 #### From [ESSENTIAL KAOS Public Repository](https://yum.kaos.st)
