+++ go.mod
@@ -3,11 +3,12 @@ module github.com/essentialkaos/yo
 go 1.19
 
 require (
+	github.com/essentialkaos/depsy v1.0.0
 	github.com/essentialkaos/ek/v12 v12.60.1
-	github.com/essentialkaos/go-simpleyaml/v2 v2.1.3
+	github.com/essentialkaos/go-simpleyaml/v2 v2.1.4
 )
 
 require (
-	golang.org/x/sys v0.5.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
